refactor: split EnrichedImage
